# extensions/a2a/reference-receiver/src/server.js
#!/usr/bin/env node
/*
 * Copyright contributors to the aiagent-portable-authorization project.
 * SPDX-License-Identifier: MIT
 * This code is generated by AI, edited by humans
 */

import http from "node:http";
import crypto from "node:crypto";
import fs from "node:fs";
import path from "node:path";
import { fileURLToPath } from "node:url";

import { createPortAuthReceiver, defaultRequestContextMapper } from "portauth-a2a-receiver";
import { A2A_MEDIA_TYPE } from "portauth-a2a";
import { createAgentCard } from "./agent-card.js";

const DEFAULT_MAX_BODY_BYTES = 512 * 1024;
const RECEIVER_AUTH_MODES = new Set(["none", "bearer"]);
const BEARER_REALM = "PortAuth A2A Reference Receiver";
const OPENAPI_PATH = path.resolve(fileURLToPath(import.meta.url), "../../../schemas/openapi.json");

// Load the OpenAPI contract once. Returns null if the file is missing so the
// /openapi.json route degrades to 404 rather than failing the whole server.
export function loadOpenApiDocument(filePath = OPENAPI_PATH) {
  try {
    return JSON.parse(fs.readFileSync(filePath, "utf8"));
  } catch {
    return null;
  }
}

export function createReferenceReceiverServer({
  engineUrl = process.env.ENGINE_URL || "http://localhost:3200",
  receiverUrl = process.env.A2A_RECEIVER_URL || `http://localhost:${process.env.PORT || "3300"}`,
  authMode = process.env.A2A_RECEIVER_AUTH_MODE || "none",
  bearerToken = process.env.A2A_RECEIVER_BEARER_TOKEN || null,
  engineAuthToken = process.env.A2A_ENGINE_AUTH_TOKEN || null,
  enableTaskLookup = process.env.A2A_ENABLE_TASK_LOOKUP === "true",
  taskReadToken = process.env.A2A_TASK_READ_TOKEN || null,
  enableMetrics = process.env.A2A_ENABLE_METRICS === "true",
  metricsToken = process.env.A2A_METRICS_TOKEN || null,
  contextMapper = defaultRequestContextMapper,
  requireProofOfPossession = resolvePopPolicy(),
  agentCardName = process.env.A2A_RECEIVER_NAME,
  agentCardDescription = process.env.A2A_RECEIVER_DESCRIPTION,
  fetchImpl = globalThis.fetch,
  logger = console,
  metrics = createMetricsCollector(),
  onEvent,
  maxBodyBytes = Number(process.env.A2A_MAX_BODY_BYTES || DEFAULT_MAX_BODY_BYTES),
} = {}) {
  const receiverAuth = normalizeReceiverAuthConfig({ authMode, bearerToken });
  const agentCard = createAgentCard({ receiverUrl, name: agentCardName, description: agentCardDescription });
  const openApiDocument = loadOpenApiDocument();
  const handleEvent = (event) => {
    metrics?.record?.(event);
    logger.info?.({ msg: "a2a.event", ...event });
    onEvent?.(event);
  };
  const portauth = createPortAuthReceiver({
    engineUrl,
    engineAuthToken,
    onEvent: handleEvent,
    fetchImpl,
    logger,
    mapRequestContext: contextMapper,
    requireProofOfPossession,
  });

  const server = http.createServer(async (req, res) => {
    try {
      const url = new URL(req.url || "/", receiverUrl);
      if (req.method === "GET" && url.pathname === "/health") {
        return sendJson(res, 200, { ok: true });
      }
      if (req.method === "GET" && url.pathname === "/openapi.json") {
        return openApiDocument
          ? sendJson(res, 200, openApiDocument)
          : sendJson(res, 404, { error: { code: "NOT_FOUND", message: "OpenAPI document unavailable" } });
      }
      if (req.method === "GET" && url.pathname === "/ready") {
        const readiness = await checkReadiness({ engineUrl, taskStore: portauth.taskStore, fetchImpl });
        return sendJson(res, readiness.ready ? 200 : 503, readiness, { "Content-Type": A2A_MEDIA_TYPE });
      }
      if (req.method === "GET" && url.pathname === "/metrics") {
        if (!enableMetrics) return sendJson(res, 404, { error: { code: "NOT_FOUND", message: "Metrics are disabled" } });
        if (metricsToken && !isAuthorizedBearer(req.headers.authorization, metricsToken)) {
          return sendJson(res, 401, { error: { code: "UNAUTHENTICATED", message: "Metrics require a bearer token" } });
        }
        return sendText(res, 200, metrics.render(), { "Content-Type": "text/plain; version=0.0.4" });
      }
      if (req.method === "GET" && url.pathname === "/.well-known/agent-card.json") {
        return sendJson(res, 200, agentCard, { "Content-Type": "application/json" });
      }
      if (req.method === "GET" && url.pathname === "/tasks") {
        if (!enableTaskLookup) return sendJson(res, 404, { error: { code: "NOT_FOUND", message: "Task lookup is disabled" } }, { "Content-Type": A2A_MEDIA_TYPE });
        if (!authorizeTaskRead(req, taskReadToken)) return sendJson(res, 401, { error: { code: "UNAUTHENTICATED", message: "Task lookup requires a bearer token" } }, { "Content-Type": A2A_MEDIA_TYPE });
        const records = await portauth.taskStore.list();
        return sendJson(res, 200, {
          tasks: records.map((record) => record.task).filter(Boolean),
          records,
        }, { "Content-Type": A2A_MEDIA_TYPE });
      }
      if (req.method === "GET" && url.pathname.startsWith("/tasks/")) {
        if (!enableTaskLookup) return sendJson(res, 404, { error: { code: "NOT_FOUND", message: "Task lookup is disabled" } }, { "Content-Type": A2A_MEDIA_TYPE });
        if (!authorizeTaskRead(req, taskReadToken)) return sendJson(res, 401, { error: { code: "UNAUTHENTICATED", message: "Task lookup requires a bearer token" } }, { "Content-Type": A2A_MEDIA_TYPE });
        const taskId = decodeURIComponent(url.pathname.slice("/tasks/".length));
        const record = await portauth.taskStore.get(taskId)
          || await portauth.taskStore.getByTaskId?.(taskId);
        return record
          ? sendJson(res, 200, { task: record.task, record }, { "Content-Type": A2A_MEDIA_TYPE })
          : sendJson(res, 404, { error: { code: "NOT_FOUND", message: `Task ${taskId} not found` } }, { "Content-Type": A2A_MEDIA_TYPE });
      }
      if (req.method === "POST" && url.pathname === "/message:send") {
        const endpointAuth = authorizeReceiverRequest(req, receiverAuth);
        if (!endpointAuth.ok) {
          metrics?.record?.({ type: "rejected", reason: "PORTAUTH_RECEIVER_UNAUTHENTICATED", status: 401 });
          return sendJson(res, endpointAuth.status, endpointAuth.body, endpointAuth.headers);
        }
        const body = await readJsonBody(req, maxBodyBytes);
        const result = await portauth.handleSendMessage({ headers: req.headers, body });
        return sendJson(res, result.status, result.body, result.headers);
      }
      return sendJson(res, 404, { error: { code: "NOT_FOUND", message: "Route not found" } });
    } catch (err) {
      logger.error?.({ msg: "A2A reference receiver error", error: err.message });
      const status = err.status || 500;
      const code = status === 413 ? "RESOURCE_EXHAUSTED" : "INTERNAL";
      const message = status === 413 ? "Request body too large" : "Reference receiver failed unexpectedly";
      return sendJson(res, status, { error: { code, message } }, { "Content-Type": A2A_MEDIA_TYPE });
    }
  });

  return { server, agentCard, taskStore: portauth.taskStore, metrics };
}

// Build a mandatory proof-of-possession policy from the environment:
//   A2A_REQUIRE_POP=true         -> require PoP for every request
//   A2A_POP_REQUIRED_ACTIONS=a,b -> require PoP only for those requestContext.action values
// Returns false (no requirement), true (all requests), or a predicate.
export function resolvePopPolicy(env = process.env) {
  if (String(env.A2A_REQUIRE_POP || "").trim().toLowerCase() === "true") return true;
  const actions = (env.A2A_POP_REQUIRED_ACTIONS || "")
    .split(",")
    .map((item) => item.trim())
    .filter(Boolean);
  if (actions.length === 0) return false;
  const required = new Set(actions);
  return ({ requestContext }) => required.has(requestContext?.action);
}

export function normalizeReceiverAuthConfig({ authMode = "none", bearerToken = null } = {}) {
  const mode = String(authMode || "none").trim().toLowerCase();
  if (!RECEIVER_AUTH_MODES.has(mode)) {
    throw new Error("A2A_RECEIVER_AUTH_MODE must be one of: none, bearer");
  }
  if (mode === "bearer" && (typeof bearerToken !== "string" || bearerToken.length === 0)) {
    throw new Error("A2A_RECEIVER_BEARER_TOKEN is required when A2A_RECEIVER_AUTH_MODE=bearer");
  }
  return { authMode: mode, bearerToken };
}

export function authorizeReceiverRequest(req, { authMode = "none", bearerToken = null } = {}) {
  const config = normalizeReceiverAuthConfig({ authMode, bearerToken });
  if (config.authMode === "none") return { ok: true };
  if (config.authMode === "bearer" && isAuthorizedBearer(req.headers.authorization, config.bearerToken)) {
    return { ok: true };
  }
  return {
    ok: false,
    status: 401,
    body: {
      error: {
        code: "UNAUTHENTICATED",
        message: "Receiver endpoint requires a bearer token",
      },
    },
    headers: {
      "Content-Type": A2A_MEDIA_TYPE,
      "WWW-Authenticate": `Bearer realm="${BEARER_REALM}"`,
    },
  };
}

export function isAuthorizedBearer(authHeader, expectedToken) {
  if (typeof expectedToken !== "string" || expectedToken.length === 0) return false;
  const header = firstHeader(authHeader);
  if (typeof header !== "string") return false;
  const match = header.match(/^Bearer\s+(.+)$/i);
  if (!match) return false;
  const suppliedToken = match[1].trim();
  if (!suppliedToken) return false;
  return timingSafeStringEqual(suppliedToken, expectedToken);
}

function firstHeader(value) {
  return Array.isArray(value) ? value[0] : value;
}

function timingSafeStringEqual(actual, expected) {
  const actualDigest = crypto.createHash("sha256").update(String(actual), "utf8").digest();
  const expectedDigest = crypto.createHash("sha256").update(String(expected), "utf8").digest();
  return crypto.timingSafeEqual(actualDigest, expectedDigest);
}

function authorizeTaskRead(req, token) {
  if (!token) return false;
  return isAuthorizedBearer(req.headers.authorization, token);
}

// ---------------------------------------------------------------------------
// Production secret handling
// ---------------------------------------------------------------------------

// Reject development placeholders so they cannot reach a production deployment.
export function isPlaceholderSecret(value) {
  if (typeof value !== "string" || value.length === 0) return true;
  return /change[-_]?me|placeholder|^changeme$|^secret$|^password$|^test$|^example|^dev-/i.test(value.trim());
}

// Resolve and validate the receiver's secret-bearing configuration from an env
// map. Fails closed: any missing or placeholder secret is an error unless
// A2A_DEV_INSECURE=true. Returns every problem at once rather than the first.
export function resolveReceiverSecrets(env = process.env) {
  const devInsecure = env.A2A_DEV_INSECURE === "true";
  const errors = [];
  const warnings = [];

  const authMode = String(env.A2A_RECEIVER_AUTH_MODE || "none").trim().toLowerCase();
  const bearerToken = env.A2A_RECEIVER_BEARER_TOKEN || null;
  const taskReadToken = env.A2A_TASK_READ_TOKEN || null;
  const engineAuthToken = env.A2A_ENGINE_AUTH_TOKEN || null;
  const metricsToken = env.A2A_METRICS_TOKEN || null;
  const enableTaskLookup = env.A2A_ENABLE_TASK_LOOKUP === "true";
  const enableMetrics = env.A2A_ENABLE_METRICS === "true";

  if (!RECEIVER_AUTH_MODES.has(authMode)) {
    errors.push("A2A_RECEIVER_AUTH_MODE must be one of: none, bearer");
  }

  const requireSecret = (name, value, condition) => {
    if (!condition) return;
    if (value == null) {
      errors.push(`${name} is required`);
    } else if (isPlaceholderSecret(value)) {
      errors.push(`${name} looks like a development placeholder; set a real secret`);
    }
  };

  requireSecret("A2A_RECEIVER_BEARER_TOKEN", bearerToken, authMode === "bearer");
  requireSecret("A2A_TASK_READ_TOKEN", taskReadToken, enableTaskLookup);
  requireSecret("A2A_METRICS_TOKEN", metricsToken, enableMetrics && metricsToken != null);

  // Any provided secret must not be a placeholder, even if not strictly required.
  for (const [name, value] of [
    ["A2A_ENGINE_AUTH_TOKEN", engineAuthToken],
    ["A2A_RECEIVER_BEARER_TOKEN", bearerToken],
    ["A2A_TASK_READ_TOKEN", taskReadToken],
    ["A2A_METRICS_TOKEN", metricsToken],
  ]) {
    if (value != null && isPlaceholderSecret(value)) {
      errors.push(`${name} looks like a development placeholder; set a real secret`);
    }
  }

  if (authMode === "none") {
    warnings.push("A2A_RECEIVER_AUTH_MODE=none: receiver /message:send is unauthenticated. Front it with mTLS or a gateway.");
  }
  if (!engineAuthToken) {
    warnings.push("A2A_ENGINE_AUTH_TOKEN is unset: receiver -> engine /evaluate calls are unauthenticated.");
  }

  const ok = errors.length === 0 || devInsecure;
  return {
    ok,
    devInsecure,
    errors: [...new Set(errors)],
    warnings: [...new Set(warnings)],
    config: { authMode, bearerToken, taskReadToken, engineAuthToken, metricsToken, enableTaskLookup, enableMetrics },
  };
}

// Throw on fatal secret problems (unless dev-insecure); always surface warnings.
export function assertReceiverSecrets(env = process.env, logger = console) {
  const result = resolveReceiverSecrets(env);
  for (const warning of result.warnings) logger.warn?.({ msg: "a2a.secret.warning", warning });
  if (!result.ok) {
    const detail = result.errors.join("; ");
    throw new Error(`Refusing to start with insecure secrets: ${detail}. Set real secrets or A2A_DEV_INSECURE=true for local use.`);
  }
  if (result.devInsecure && result.errors.length > 0) {
    logger.warn?.({ msg: "a2a.secret.dev_insecure", errors: result.errors });
  }
  return result;
}

// ---------------------------------------------------------------------------
// Operational monitoring
// ---------------------------------------------------------------------------

// Readiness probe: the receiver is only ready when its hard dependencies
// (PortAuth engine and the task store) are reachable.
export async function checkReadiness({ engineUrl, taskStore, fetchImpl = globalThis.fetch, timeoutMs = 2000 } = {}) {
  const engine = await probe(async () => {
    const controller = new AbortController();
    const timer = setTimeout(() => controller.abort(), timeoutMs);
    try {
      const resp = await fetchImpl(new URL("health", ensureTrailingSlash(engineUrl)), { signal: controller.signal });
      return resp.ok;
    } finally {
      clearTimeout(timer);
    }
  });
  const store = await probe(async () => {
    await taskStore.list({ limit: 1 });
    return true;
  });
  return {
    ready: engine.ok && store.ok,
    checks: { engine, store },
  };
}

async function probe(fn) {
  try {
    const ok = await fn();
    return { ok: Boolean(ok) };
  } catch (err) {
    return { ok: false, error: err.message };
  }
}

// Minimal in-process metrics collector that renders Prometheus text. Consumes
// the receiver guard's lifecycle events. Real deployments would swap this for
// an OpenTelemetry exporter via the same onEvent seam.
export function createMetricsCollector() {
  const decisions = new Map();
  const rejections = new Map();
  let replays = 0;
  const persistenceFailures = new Map();
  const engineLatency = { sum: 0, count: 0 };
  const totalLatency = { sum: 0, count: 0 };

  return {
    record(event = {}) {
      if (!event || typeof event !== "object") return;
      if (typeof event.engineDurationMs === "number") {
        engineLatency.sum += event.engineDurationMs;
        engineLatency.count += 1;
      }
      if (typeof event.durationMs === "number") {
        totalLatency.sum += event.durationMs;
        totalLatency.count += 1;
      }
      switch (event.type) {
        case "decision":
          increment(decisions, event.decision || "unknown");
          break;
        case "rejected":
          increment(rejections, event.reason || "unknown");
          break;
        case "replay":
          replays += 1;
          break;
        case "persistence_failed":
          increment(persistenceFailures, event.phase || "unknown");
          break;
        default:
          break;
      }
    },
    snapshot() {
      return {
        decisions: Object.fromEntries(decisions),
        rejections: Object.fromEntries(rejections),
        replays,
        persistenceFailures: Object.fromEntries(persistenceFailures),
        engineLatency: { ...engineLatency },
        totalLatency: { ...totalLatency },
      };
    },
    render() {
      const lines = [];
      lines.push("# HELP portauth_a2a_decisions_total Authorization decisions by outcome.");
      lines.push("# TYPE portauth_a2a_decisions_total counter");
      for (const [decision, value] of decisions) {
        lines.push(`portauth_a2a_decisions_total{decision="${escapeLabel(decision)}"} ${value}`);
      }
      lines.push("# HELP portauth_a2a_rejected_total Requests rejected before engine evaluation.");
      lines.push("# TYPE portauth_a2a_rejected_total counter");
      for (const [reason, value] of rejections) {
        lines.push(`portauth_a2a_rejected_total{reason="${escapeLabel(reason)}"} ${value}`);
      }
      lines.push("# HELP portauth_a2a_replays_total Idempotent replays served from the task store.");
      lines.push("# TYPE portauth_a2a_replays_total counter");
      lines.push(`portauth_a2a_replays_total ${replays}`);
      lines.push("# HELP portauth_a2a_persistence_failures_total Task persistence failures by phase.");
      lines.push("# TYPE portauth_a2a_persistence_failures_total counter");
      for (const [phase, value] of persistenceFailures) {
        lines.push(`portauth_a2a_persistence_failures_total{phase="${escapeLabel(phase)}"} ${value}`);
      }
      lines.push("# HELP portauth_a2a_engine_call_duration_ms_sum Total engine call latency.");
      lines.push("# TYPE portauth_a2a_engine_call_duration_ms_sum counter");
      lines.push(`portauth_a2a_engine_call_duration_ms_sum ${engineLatency.sum}`);
      lines.push(`portauth_a2a_engine_call_duration_ms_count ${engineLatency.count}`);
      lines.push("# HELP portauth_a2a_request_duration_ms_sum Total end-to-end request latency.");
      lines.push("# TYPE portauth_a2a_request_duration_ms_sum counter");
      lines.push(`portauth_a2a_request_duration_ms_sum ${totalLatency.sum}`);
      lines.push(`portauth_a2a_request_duration_ms_count ${totalLatency.count}`);
      return `${lines.join("\n")}\n`;
    },
  };
}

function increment(map, key) {
  map.set(key, (map.get(key) || 0) + 1);
}

function escapeLabel(value) {
  return String(value).replace(/\\/g, "\\\\").replace(/"/g, '\\"');
}

// ---------------------------------------------------------------------------
// Structured, redacting logger
// ---------------------------------------------------------------------------

const REDACT_KEYS = new Set(["authorization", "vc", "credential", "credentials", "password"]);

// JSON logger that drops secret-bearing fields. Defends SPEC §8.1 (never log
// raw VCs) and keeps tokens out of logs even if an event accidentally carries one.
export function createStructuredLogger({ stream = process.stdout, errorStream = process.stderr } = {}) {
  const write = (target, level, payload) => {
    const line = JSON.stringify({ ts: new Date().toISOString(), level, ...redact(payload) });
    target.write(`${line}\n`);
  };
  return {
    info: (payload) => write(stream, "info", payload),
    warn: (payload) => write(stream, "warn", payload),
    error: (payload) => write(errorStream, "error", payload),
  };
}

export function redact(value) {
  if (Array.isArray(value)) return value.map(redact);
  if (value && typeof value === "object") {
    const out = {};
    for (const [key, val] of Object.entries(value)) {
      const lower = key.toLowerCase();
      if (REDACT_KEYS.has(lower) || lower.endsWith("token") || lower.endsWith("secret")) {
        out[key] = "[redacted]";
      } else {
        out[key] = redact(val);
      }
    }
    return out;
  }
  return value;
}

export async function startReferenceReceiver(options = {}) {
  const port = Number(options.port || process.env.PORT || 3300);
  const host = options.host || process.env.HOST || "0.0.0.0";
  const { server, agentCard, taskStore, metrics } = createReferenceReceiverServer(options);
  await new Promise((resolve) => server.listen(port, host, resolve));
  return { server, agentCard, taskStore, metrics, port, host };
}

function sendJson(res, status, body, headers = {}) {
  const payload = JSON.stringify(body, null, 2);
  res.writeHead(status, {
    "Content-Type": "application/json",
    "Content-Length": Buffer.byteLength(payload),
    ...headers,
  });
  res.end(payload);
}

function sendText(res, status, text, headers = {}) {
  res.writeHead(status, {
    "Content-Type": "text/plain; charset=utf-8",
    "Content-Length": Buffer.byteLength(text),
    ...headers,
  });
  res.end(text);
}

function ensureTrailingSlash(value) {
  const text = String(value);
  return text.endsWith("/") ? text : `${text}/`;
}

async function readJsonBody(req, maxBodyBytes) {
  const chunks = [];
  let size = 0;
  for await (const chunk of req) {
    size += chunk.length;
    if (size > maxBodyBytes) {
      throw Object.assign(new Error("request body too large"), { status: 413 });
    }
    chunks.push(chunk);
  }
  const text = Buffer.concat(chunks).toString("utf8");
  return text ? JSON.parse(text) : {};
}

// Resolve an optional named context mapper for the CLI. The receiver is
// domain-neutral by default; setting A2A_CONTEXT_MAPPER=bill-payment loads the
// bundled banking example to demonstrate dependency injection of a domain mapper.
export async function resolveCliContextMapper(env = process.env) {
  const name = (env.A2A_CONTEXT_MAPPER || "").trim().toLowerCase();
  if (!name || name === "default") return undefined;
  if (name === "bill-payment") {
    const { billPaymentContextMapper } = await import("../examples/bill-payment-mapper.js");
    return billPaymentContextMapper;
  }
  throw new Error(`Unknown A2A_CONTEXT_MAPPER '${name}'. Use 'default' or 'bill-payment', or inject contextMapper programmatically.`);
}

const isCli = process.argv[1] === fileURLToPath(import.meta.url);
if (isCli) {
  const logger = createStructuredLogger();
  assertReceiverSecrets(process.env, logger);
  const contextMapper = await resolveCliContextMapper(process.env);
  const { port, host } = await startReferenceReceiver({ logger, contextMapper });
  logger.info?.({ msg: "a2a.receiver.listening", url: `http://${host}:${port}`, contextMapper: process.env.A2A_CONTEXT_MAPPER || "default" });
}